+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 ### PaymentsUI
 * [Fixed] Fixed an issue with `STPPaymentCardTextField` where the `paymentCardTextFieldDidEndEditing` delegate method was not called.
 
+### PaymentSheet
+* [Fixed] Fixed some payment method icons not updating to use the latest assets.
+
 ## 23.18.0 2023-10-23
 ### PaymentSheet
 * [Added] Saved SEPA payment methods are now displayed to the customer for reuse, similar to saved cards.